% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesean_to_polar.R
\name{cartesianToPolar}
\alias{cartesianToPolar}
\title{Get polar coordinates froma a cartesean grid}
\usage{
cartesianToPolar(xy.grid)
}
\arguments{
\item{xy.grid}{A list containing the 'x' and 'y' matrices for the cartesian cooridantes
returns a list of 'theta' the  radian angle (range, -pi to pi) and corresponding 'distance' matrices}
}
\description{
Get polar coordinates froma a cartesean grid
}
\examples{
xycoord = meshgrid(x = 1:5, y = 1:5)
cartesianToPolar(xycoord)

}
