% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional_psd.R
\name{directional.spectrum}
\alias{directional.spectrum}
\title{Computes directional, radially averaged spectral density of a matrix
Used the base::fft() function to get fourier coefficients of the input matrix}
\usage{
directional.spectrum(x, alpha, offset = 10, res = 1, Log = TRUE,
  normalized = FALSE, correct.mean = FALSE, pad.values = NaN,
  plot = TRUE, ...)
}
\arguments{
\item{x}{a matrix}

\item{alpha}{the angle at which the rspectrum is calculated (0 - 90)}

\item{offset}{offset to (alpha + - offset is be used to calculate teh rspectrum)}

\item{res}{resolution of the image represented by the matrix)}

\item{Log}{Logical, whether the fft(x) values should be log-transformed, default  = TRUE}

\item{normalized}{Logical, whetther the fft (x) values should be normalized, default = FALSE}

\item{correct.mean}{Logical, whether the input x should be subject to mean correction, default = FALSE}

\item{pad.values}{In case x is not a square, the values to apply to the new cells added in the periphery, default is NaN}

\item{plot}{Logical, whether the respectrum output should be ploted, default = TRUE}
}
\description{
Computes directional, radially averaged spectral density of a matrix
Used the base::fft() function to get fourier coefficients of the input matrix
}
\examples{
data(patches2)
directional.spectrum(x = patches2, alpha = 90, offset = 10, Log = T)
}
