% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftshift.R
\name{fftshift}
\alias{fftshift}
\title{Moves zero frequency elements (of fft() output) to the center of the matrix
this makes easier to visualize the fft output.}
\usage{
fftshift(x)
}
\arguments{
\item{x}{input matrix}
}
\description{
Moves zero frequency elements (of fft() output) to the center of the matrix
this makes easier to visualize the fft output.
}
\examples{
a = matrix(1:25, 5, 5)
fftshift(a)
}
