% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshgrid.R
\name{meshgrid}
\alias{meshgrid}
\title{Get a mesh grid defined by the ranges of input parameters x and y}
\usage{
meshgrid(x, y)
}
\arguments{
\item{x}{a vector of numerica values}

\item{y}{a vector of numeric values}
}
\description{
Get a mesh grid defined by the ranges of input parameters x and y
}
\examples{
x = 1:5
y = 1:5
meshgrid(x, y)


}
