% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_spectral_density.R
\name{radial.spectrum}
\alias{radial.spectrum}
\title{Computes radially averaged spectral density of a matrix}
\usage{
radial.spectrum(x, res = 1, Log = TRUE, normalized = FALSE,
  correct.mean = FALSE, pad.values = NaN, plot = TRUE, ...)
}
\arguments{
\item{x}{a matrix}

\item{res}{resolution of the image represented by the matrix)}

\item{Log}{Logical, whether the fft(x) values should be log-transformed, default  = TRUE}

\item{normalized}{Logical, whetther the fft (x) values should be normalized, default = FALSE}

\item{correct.mean}{Logical, whether the input x should be subject to mean correction, default = FALSE}

\item{pad.values}{In case x is not a square, the values to apply to the new cells added in the periphery, default is NaN}

\item{plot}{Logical, whether the respectrum output should be ploted, default = TRUE}
}
\description{
Computes radially averaged spectral density of a matrix
}
\examples{
data(patches)
radial.spectrum(x = patches, Log = T)
}
