% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect2sqMatrix.R
\name{rect2sqMatrix}
\alias{rect2sqMatrix}
\title{Convert a rectangular matrix object to a square on by addign rows/columns}
\usage{
rect2sqMatrix(x, pad = NULL, trim = NULL, pad.values = NaN, ...)
}
\arguments{
\item{x}{input matrix}

\item{pad.values}{values to be applied to the new rows/columns. devault = NaN}
}
\description{
Convert a rectangular matrix object to a square on by addign rows/columns
}
\examples{

x = matrix(0, 6, 3)
make.square(x)
}
